RSA FIX
After sometime of messing around with the code, I was able to collaborate with my professor in debugging my program a little and figuring out why it overall wasn't working. In this commit I submit a new version of my stoi and itos functions, as well as an update to the getPrimes method as I realized that sometimes my p * q = n was too small and when the n was too small it would make the modulus between 0-small n which was under the string_as_an_int which would have made it impossible to turn the string_as_int into a cipher back into the string_as_int with a value lower than expected. Fixing these two bugs, I was able to make the program work!

The biggest note to the program is that it only works for strings of length 3 (exp: "cow")...length 4 if you are counting it being null terminated as "[c] [o] [w] [\0]"

I will attempt to fix little things in the next commit: Making getPrimes receive 2 random numbers instead of needing two different functions...